backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
python-multipart==0.0.12
pyjwt[crypto]==2.9.0
passlib[bcrypt]==1.7.4
bcrypt>=3.2,<4
python-dotenv==1.0.1
pillow>=10.4,<12
pydantic[email]>=2.10,<3
pydantic-settings>=2.6,<3
pandas>=2.2,<3
python-dateutil==2.9.0
alembic==1.14.0
